Add node for screen-space shaders
I called the node `ScreenSpaceMesh` and it only needs to have a shader assigned to it (example shaders provided). It just needs to be placed within in the scene tree (doesn't need to be a child of a `Camera3D`).

Multiple shader passes are also possible since the node just takes a normal `ShaderMaterial`. It's kinda op ngl


Former-commit-id: 0ef70d4c3006cabf68f6f2504d01feac43aa0c2f