Add film grain overlay
I thought it might give the game's visuals a bit more character (I think I made it subtle enough to not be distracting).

The effect is always enabled with a max opacity of 5.5%. I believe it is performant enough for any device, but we can make a setting to disable it if needed.